api/requirements.txt
psycopg2==2.9.10
Django==5.1.2
djangorestframework==3.15.2
requests==2.32.3
python-dotenv==1.0.1
